move tests to a different file
